# Speed up JSONRenderer.extract_included