@mblayman 2018-02-15
@amw I'm definitely open to this fix. I think the warning you described sounds like acceptable risk to me. Looks like there is an isort error in the CI build (I think it's an ASCII sorting issue thing). I'm happy to review once Travis is green.