@amw 2018-02-15
I've sent a fix for `isort`, but Travis is still complaining – I don't see about what though – all the logs seem empty.