@mblayman 2018-02-16
Looks like you fixed isort. Now you have a flake8 error.
```
$ flake8
./rest_framework_json_api/renderers.py:318:101: E501 line too long (103 > 100 characters)
./rest_framework_json_api/renderers.py:401:33: E126 continuation line over-indented for hanging indent
./rest_framework_json_api/renderers.py:406:29: E121 continuation line under-indented for hanging indent
./rest_framework_json_api/renderers.py:424:29: E126 continuation line over-indented for hanging indent
The command "flake8" exited with 1.
```